Use Arm.TriggerAnimation in MouseInteract
There is an error in the logic that sets the 'retract' trigger for the
arm animator if you 'reach' the arm while the arm is in the retracting
animation. I could fix it in the current code, but I suspect the better
way forward would be to keep arm state in a separate Bolt graph.